# modules/apps/data-engine/data-engine-rest-impl/rest-openapi.yaml
components:
    schemas:
...
        DataRecord:
            description: https://www.schema.org/DataRecord
            properties:
                dataRecordCollectionId:
                    format: int64
                    type: integer
                dataRecordValues:
                    additionalProperties:
                        type: object
                    type: object
                id:
                    format: int64
                    type: integer
...
info:
    description: ""
    title: ""
    version: v1.0
openapi: 3.0.1
paths:
...
    "/data-record-collections/{dataRecordCollectionId}/data-records":
        get:
            operationId: getDataRecordCollectionDataRecordsPage
            parameters:
                - in: path
                  name: dataRecordCollectionId
                  required: true
                  schema:
                      format: int64
                      type: integer
                - in: query
                  name: fieldNames
                  schema:
                      items:
                          type: string
                      type: array
                - in: query
                  name: page
                  schema:
                      type: integer
                - in: query
                  name: pageSize
                  schema:
                      type: integer
            responses:
                200:
                    content:
                        application/json:
                            schema:
                                items:
                                    $ref: "#/components/schemas/DataRecord"
                                type: array
                        application/xml:
                            schema:
                                items:
                                    $ref: "#/components/schemas/DataRecord"
                                type: array
                    description: ""
            tags: ["DataRecord"]
        post:
            operationId: postDataRecordCollectionDataRecord
            parameters:
                - in: path
                  name: dataRecordCollectionId
                  required: true
                  schema:
                      format: int64
                      type: integer
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/DataRecord"
                    application/xml:
                        schema:
                            $ref: "#/components/schemas/DataRecord"
            responses:
                200:
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/DataRecord"
                        application/xml:
                            schema:
                                $ref: "#/components/schemas/DataRecord"
                    description: ""
            tags: ["DataRecord"]
    "/data-record-collections/{dataRecordCollectionId}/data-records/export":
        get:
            operationId: getDataRecordCollectionDataRecordExport
            parameters:
                - in: path
                  name: dataRecordCollectionId
                  required: true
                  schema:
                      format: int64
                      type: integer
                - in: query
                  name: fieldNames
                  schema:
                      items:
                          type: string
                      type: array
                - in: query
                  name: page
                  schema:
                      type: integer
                - in: query
                  name: pageSize
                  schema:
                      type: integer
            responses:
                204:
                    content:
                        application/json:
                            schema:
                                type: string
                    description: ""
            tags: ["DataRecord"]
    "/data-records/{dataRecordId}":
        delete:
            operationId: deleteDataRecord
            parameters:
                - in: path
                  name: dataRecordId
                  required: true
                  schema:
                      format: int64
                      type: integer
            responses:
                204:
                    content:
                        application/json: {}
                    description: ""
            tags: ["DataRecord"]
        get:
            operationId: getDataRecord
            parameters:
                - in: path
                  name: dataRecordId
                  required: true
                  schema:
                      format: int64
                      type: integer
            responses:
                200:
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/DataRecord"
                        application/xml:
                            schema:
                                $ref: "#/components/schemas/DataRecord"
                    description: ""
            tags: ["DataRecord"]
        put:
            operationId: putDataRecord
            parameters:
                - in: path
                  name: dataRecordId
                  required: true
                  schema:
                      format: int64
                      type: integer
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/DataRecord"
                    application/xml:
                        schema:
                            $ref: "#/components/schemas/DataRecord"
            responses:
                200:
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/DataRecord"
                        application/xml:
                            schema:
                                $ref: "#/components/schemas/DataRecord"
                    description: ""
            tags: ["DataRecord"]
...